fixed the password system
Had some minor bugs like I forgot to tell the the client when a password was determined necessary. Also I was checking for unathorized users by seeing if their auth level was <= 3 which would include everybody, so i lowered it to 2. The password prompt door should be sending power to the same side as the gate, so I removed side2 completely. I also had to expand the box to include the inner walls of the gate; players weren't being detected when at the edge of the room, potentially due to some rounding error.